refactor: перенести навигационный хелпер в context.symbol.SourceDefinedSymbolLinks
utils/ — листовой пакет и не должен зависеть от доменных типов (context.symbol),
поэтому хелпер построения таргета uri#L<line>,<col> переехал из
utils.SourceSymbolLinks в context.symbol.SourceDefinedSymbolLinks и снова
принимает SourceDefinedSymbol (чистый однопараметрический API). Вызовы в hover
(SeeReferenceHyperlinks) и documentlink (SeeReferenceDocumentLinkSupplier)
обновлены.
